Move threadPoolCallbackRunner to the "Common" folder.
