+++ src/Disks/IO/ThreadPoolReader.h
@@ -3,7 +3,7 @@
 #include <memory>
 #include <IO/AsynchronousReader.h>
 #include <Common/ThreadPool_fwd.h>
-#include <Interpreters/threadPoolCallbackRunner.h>
+#include <Common/threadPoolCallbackRunner.h>
 
 
 namespace DB
