+++ src/IO/S3/copyS3File.h
@@ -5,7 +5,7 @@
 #if USE_AWS_S3
 
 #include <Storages/StorageS3Settings.h>
-#include <Interpreters/threadPoolCallbackRunner.h>
+#include <Common/threadPoolCallbackRunner.h>
 #include <IO/S3/BlobStorageLogWriter.h>
 #include <base/types.h>
 #include <functional>
